refactor(functions): share worker peer test harness
