+++ packages/cli/test/functions-rtdb/child.integration.test.ts
@@ -4,9 +4,6 @@ import { existsSync, mkdirSync, mkdtempSync, symlinkSync, writeFileSync } from '
 import { tmpdir } from 'node:os';
 import { join, resolve } from 'node:path';
 import { pathToFileURL } from 'node:url';
-import WebSocket from 'ws';
-import { createMemoryBackend, initializeSandbox } from 'pyric/sandbox';
-import { getFirestore } from 'pyric/firestore';
 import {
   spawnFunctionsRtdbChild,
   type FunctionsRtdbChildEvent,
@@ -17,22 +14,12 @@ import {
   registerModuleUrl,
 } from '../../src/cli/dev-runner.js';
 import { startServe, type ServeRuntime } from '../../src/cli/serve.js';
-import {
-  isBridgeMessage,
-  WORKER_RELAY_CAPABILITY,
-  type BridgeMessage,
-} from '../../src/bridge/protocol.js';
 import { connectRemoteSandbox, type RemoteSandbox } from '../../src/remote/index.js';
 import { silentServeLogger } from '../../src/serve/server.js';
 import {
-  handleMessage,
-  type HostCtx,
-  type PortLike,
-} from '../../src/serve/worker/host.js';
-import type {
-  InboundMessage,
-  OutboundMessage,
-} from '../../src/serve/worker/protocol.js';
+  connectFunctionsWorkerPeer,
+  createFunctionsWorkerHostCtx,
+} from './worker-peer.js';
 
 const cliRoot = resolve(import.meta.dir, '../..');
 const repoRoot = resolve(cliRoot, '../..');
@@ -43,96 +30,6 @@ let runtime: ServeRuntime;
 let peer: { close(): Promise<void> };
 let observer: RemoteSandbox;
 let child: FunctionsRtdbChildHandle | undefined;
-
-async function makeWorkerCtx(): Promise<HostCtx> {
-  const sandbox = initializeSandbox();
-  await sandbox.enablePersistence({
-    key: `functions-child-${Math.random()}`,
-    injectedBackend: createMemoryBackend(),
-  });
-  return {
-    db: getFirestore(sandbox),
-    sandbox,
-    instanceId: 'functions-child-test',
-    subs: new Map(),
-  };
-}
-
-async function connectWorkerPeer(
-  url: string,
-  ctx: HostCtx,
-): Promise<{ close(): Promise<void> }> {
-  const ws = new WebSocket(url);
-  const port: PortLike = {
-    postMessage(raw: unknown) {
-      if (ws.readyState !== WebSocket.OPEN) return;
-      const message = raw as OutboundMessage;
-      if (message.t === 'res') {
-        const response: BridgeMessage = message.ok
-          ? { type: 'worker-res', id: message.id, ok: true, value: message.value }
-          : { type: 'worker-res', id: message.id, ok: false, error: message.error };
-        ws.send(JSON.stringify(response));
-      } else if (message.t === 'snap') {
-        ws.send(JSON.stringify({
-          type: 'worker-snap',
-          subId: message.subId,
-          value: message.value,
-        } satisfies BridgeMessage));
-      }
-    },
-  };
-
-  await new Promise<void>((resolveConnected, rejectConnected) => {
-    ws.once('open', () => {
-      ws.send(JSON.stringify({
-        type: 'hello',
-        protocol: 1,
-        tools: [],
-        sandboxId: 'functions-child-peer',
-        capabilities: [WORKER_RELAY_CAPABILITY],
-      } satisfies BridgeMessage));
-    });
-    ws.on('message', (raw) => {
-      let message: unknown;
-      try {
-        message = JSON.parse(raw.toString());
-      } catch {
-        return;
-      }
-      if (!isBridgeMessage(message)) return;
-      if (message.type === 'hello-ack') {
-        resolveConnected();
-      } else if (message.type === 'worker-op') {
-        void handleMessage(ctx, port, {
-          ...message.op,
-          t: 'op',
-          id: message.id,
-        } as InboundMessage);
-      } else if (message.type === 'worker-sub') {
-        void handleMessage(ctx, port, {
-          ...message.sub,
-          t: 'sub',
-          subId: message.subId,
-        } as InboundMessage);
-      } else if (message.type === 'worker-unsub') {
-        void handleMessage(ctx, port, {
-          t: 'unsub',
-          subId: message.subId,
-        } satisfies InboundMessage);
-      }
-    });
-    ws.once('error', rejectConnected);
-    ws.once('close', () => rejectConnected(new Error('worker peer closed before ready')));
-  });
-
-  return {
-    close: () => new Promise<void>((resolveClosed) => {
-      if (ws.readyState === WebSocket.CLOSED) return resolveClosed();
-      ws.once('close', () => resolveClosed());
-      ws.close();
-    }),
-  };
-}
 
 async function waitForValue(path: string, expected: unknown): Promise<void> {
   const deadline = Date.now() + 5_000;
@@ -191,10 +88,15 @@ exports.unsupportedUpdate = onValueUpdated(
     bridge: true,
     disableAuditLog: true,
   });
-  peer = await connectWorkerPeer(
-    `ws://127.0.0.1:${runtime.handle.port}/__pyric/sandbox`,
-    await makeWorkerCtx(),
-  );
+  const ctx = await createFunctionsWorkerHostCtx({
+    persistenceKeyPrefix: 'functions-child',
+    instanceId: 'functions-child-test',
+  });
+  peer = await connectFunctionsWorkerPeer({
+    url: `ws://127.0.0.1:${runtime.handle.port}/__pyric/sandbox`,
+    ctx,
+    sandboxId: 'functions-child-peer',
+  });
   observer = await connectRemoteSandbox({ url: runtime.handle.url });
 });
